Adjusted based on client preference

+++ database/migrations/2020_05_22_115139_create_contacts_table.php
@@ -16,10 +16,10 @@ class CreateContactsTable extends Migration
         Schema::create('contacts', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('full_name');
-            $table->bigInteger('created_by');
+            $table->unsignedBigInteger('created_by');
             $table->string('email')->nullable();
             $table->string('status');
-            $table->integer('company_id')->nullable();
+            $table->unsignedBigInteger('company_id')->nullable();
             $table->string('tags')->nullable();
             $table->string('birthday')->nullable();
             $table->text('address');
